backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# gRPC
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2

# Deep Learning & Computer Vision
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
scipy==1.12.0

# Object Detection & Segmentation
ultralytics==8.1.11
transformers==4.37.2
timm==0.9.12
supervision==0.18.0

# ONNX Runtime
onnxruntime-gpu==1.17.0
onnx==1.15.0

# LLM Integration
ollama==0.1.6
huggingface-hub==0.20.3
sentencepiece==0.1.99
accelerate==0.26.1

# Data Processing
pandas==2.2.0
imageio==2.33.1
scikit-image==0.22.0
albumentations==1.3.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
aiofiles==23.2.1
pyyaml==6.0.1
typer==0.9.0
rich==13.7.0
tenacity==8.2.3
